docker_run: Make docker stop more resilient, and stop removing container
In some instances, the docker images were stopped and removed before the
trap was hit. This meant that the test was subsequently failed because
the container no longer existed.

I've made it more resilient by looping over a list, rather than the
common shorthand. I've also stopped removing the container because the
removal means that we lose the logs too.