runner: Move path to subpath
The /var/www directory is owned by root.
In 31 we simply append a run number to the behatroot directory - i.e.
/var/www/behatdata1

In 32 we create them in a subdirectory - i.e.
/var/www/behatdata/behatrun1

When running as www-data, the user does not have access to create the
behatdata1 directory.

The previous attempted fix was to add a slash:
31  - /var/www/behatdata/1
32+ - /var/www/behatdata/behatrun1

However, the trailing slash seems to cause issues in core in some
sitautions - specifically the re-run of the test.

This fix, tested on 31, and master with a failing test changes the
directorty further such that the behat data directory is
  /var/www/behatdata/run
This makes the new run directories:
31  - /var/www/behatdata/run1
32+ - /var/www/behatdata/run/behatrun1

Both of these are acceptable as the www-data user has the required
permission to create new directories for the run in the subdirectory.