Unpin 3.141.59-mercury and allow any new 3.x release to be used
With MDL-66378 and, specially, https://github.com/moodlehq/php-webdriver/pull/1
we can continue using Chrome >= 76 because we are enforcing the
use of the legacy JsonWire protocol (instead of the default W3C one).

So it has been agreed that we'll allow any 3.x new release to be used in
order to detect problems earlier (we always can pin again if needed).