Unpin selenium 3.141.59-zinc (Chrome 79) so 89 takes the baton
We were waiting since some time ago to Chrome 89 to be released
and, because of that, we had selenium pinned to 3.141.59-zinc
version (Chrome 79).

The cause for staying sticky was this issue:
    https://bugs.chromium.org/p/chromedriver/issues/detail?id=3682&q=&can=1&sort=-id

Chrome (and chromedriver) 89 have been released and Selenium has
already build the docker images for them:

https://github.com/SeleniumHQ/docker-selenium/releases/tag/3.141.59-20210311

So now we can unpin the old version and let the machinery to use the
latest versions available. That's what this PR does.