Unpin chrome version and cross fingers
We are going brave here and go back to unpinned chrome version, because staying with
the old Chrome 79 version (3.141.59-zinc) is making things really harder and harder.
We are aware that, for headed runs, there are some zero-size errors still not fixed,
but headless ones should run ok. For reference, the problems we are aware are being
tracked (as of 14 Sep 2021) @:
  - MDL-71108 : zero-size
  - MDL-72306 : feedback