[runner] Support phpunit coverage generation using pcov
This commit adds support for coverage generation in phpunit tests using
the `pcov` php module, which is now installed on all of our current
images.

Generation is generated and saved to clover4j and crap4j file formats,
which have minimal file size implications.

It was found during testing that the phpunit formats were simply too
large, as were all other alternatives, with the cobertura output
reaching 7-8Gb.

This change is controlled by the COVERAGE environment option. When set
to the value "pcov", coverage will be generated using pcov.