Add support for specifying the DB tags (versions) to use
Added a new parameter, DBTAG, which behaviour is as follows:

- If not specified, every database will use the currently
  pinned version (if there is any) or "latest" if not. Right now,
  the versions are:
    - mariadb: 10.7
    - mysqli: 5.7
    - oracle: latest (11)
    - pgsql: 12
    - sqlsrv: 2017-latest
- If specified, then that tag will be used, say, for example:
    - mariadb: 10.6, 10.5,...
    - mysqli: 8.0,...
    - oracle: 19, 21,...
    - sqlsrv: 2019-latest,....

So, this change doesn't make any difference in current default
jobs (only pgsql has been raised to 12 to be ready for 4.1). Everything
will continue running exactly as it's now.

Only when we add support for that new DBTAG parameter in jobs (and in
nightlyjobs) is when we'll be able to run alternative versions.