Make mariadb replicas to work with 10.2 to 10.7
The code used here is 99% the same than the mysql one. But finally
I've decided to copy it all over own mariadb.d directory.

The only difference is when creating the replication database user
because the MySQL sentence:

CREATE USER 'replication'@'%' IDENTIFIED WITH mysql_native_password BY 'replication';

Doesn't work with MariaDB. In fact, since MariaDB 10.2 the
mysql_native_password is the default authentication plugin
so it doesn't need to be specified. So this works the same:

CREATE USER 'replication'@'%' IDENTIFIED BY 'replication';

A working alternative is also:

CREATE USER 'replication'@'%' IDENTIFIED WITH mysql_native_password USING PASSWORD('replication');

But I've gone with the simple one because it's future better.

As said, everything else is 100% the same, but better split that making
complicated conditionals to get both working with the same config files.