Make pgsql replicas to work with PostgreSQL 10 and up
This seems to be working for PG 10, 11, 12, 13 and 14. I've finally
left 9.6 out from the equation because it required to add some more
shell scripts to be executed conditionally. And I think we can
ignore that old versions @ CIs already (Moodle 3.9 works with PG 12
perfectly).

Note the changes here seem more complex, but basically it's because
I've moved the stuff to be similar to the mysql/mariadb one, aka:

- Create master / slave / standalone directories.
- Move as much as possible to .sql entry points.
- Keep shell scripts simple, only doing things really requiring a shell.

Other than that, just renamed some setting or remove some stuff
that was not needed (archiving, logging...) apparently.

Fixes #79.