Limit the tmpfs usage and ensure no binlogs in standalone runs
With the switch to MySQL, bin logs are enabled by default and,
for complete phpunit/behat runs they become huge, needing > 4Gb
for completing.

While this was not affecting us in the CI infrastructure, because
there are up to 32Gb available for tmpfs in the workers, we don't
need those bin logs at all.

So this commit does:

1. Limit the tmpfs to be max 1Gb (latest tests are using ~300Mb).
2. Ensure that bin logs are disabled for mysql/mariadb runs.

Note that point 2 only affects to standalone runs, of course
primary/replica runs do require the binary logs to be there
so we aren't touching them now.

Surely they will be revisited when we try to fix existing #84. But
that's another story...