Add missing assignment of latest exit code
This was causing parallel = 1 runs to end with error whenever
the reruns had passed all the tests remaining (because the
original EXITCODE variable, that failed in the initial run
was not being updated ever.

You can see that, some lines below, for parallel > 1 runs, that
are the ones we normally do @ CIs the very same assignment is
done, and because of that we had not detected this problem that
only affects to parallel = 1.

With this change, we are feeding EXITCODE correctly for them.