Mount the whole oradata directory for tmpfs
While historically it has been enough to just mount the XE/XEPDB1
directories for tmpfs, recent changes in the upstream images (switching
from unzip to 7zzs) are causing some problems with the later
failing because it's not able to set attributes on the mounted
directory.

To workaround that (haven't been able to find any switch in 7zzs
able to allow it to continue on attribute/timemodified changes),
we are going to mount with the whole oradata directory instead.

It will require (safeguard) some more memory (+2Gb) but that
shouldn't be a problem as far as we have enough memory in the
workers and also, have Oracle concurrent builds restricted to
be max 1 by worker.

For a little more information about the issue and what caused it:

https://github.com/gvenzl/oci-oracle-xe/issues/202