First cut to the new, split into multiple, ci runner
This is the very first version of the new ci runner that
splits jobs into smaller modules, available for jobs to
orchestrate them.

This includes a working version of the phpunit job, to
be used as a reference for the migration of the behat
tests and also to create some new job to experiment
with the flexibility of the system.

Note that this is not production-ready yet, just work
in progress to go trying stuff. Once the system is stable
and tested enough we'll go advancing with the implementation plan.

For more information and updates, this is the issue to follow:

https://github.com/moodlehq/moodle-ci-runner/issues/111