Add some basic tests
This covers the runner (run and lib) and the
behat and phpunit job types.

It's not exhaustive, but covers ~60% of the code
(really more, because the coverage tool does skip
all multilines).

More tests can be added later, covering various
checks and special situations that right now aren't
being tested.

Also, detected by the tests, add a few modifications
(cosmetic) to the runner and some module.