Add support to new FULLGIT env variable
With that new env variable, any git repository being copied
to the container (or mounted as volume in the future) it's possible to:
- dissociate any --reference repository completely.
- un-shallow the repo if it was cloned with --depth
- get access to all branches if it was cloned with --single-branch

Note that it defaults to disabled (empty), so none of the current
supported jobs will use that new feature.

But other, incoming, jobs, needing a full clone to be able to
play with the repository (fetch, merge, checkout, ...) will have
it enabled to ensure that the repository is complete and free
of dependencies.